pyproject.toml
[tool.poetry]
name = "architect"
version = "0.1.0"
description = ""
authors = ["BloomTech <ai@bloomtech.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
langchain-openai = "^0.1.8"
langchain-core = "^0.2.5"
python-dotenv = "^1.0.1"
langchain_experimental = "^0.0.60"
langgraph = "^0.0.66"
langchain = "^0.2.3"
fastapi = "^0.115.0"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
